Fix tst_qquickmenubar::menuPosition for all styles
The test assumed that the editMenu's insets were {0, 0}.
This caused the test to fail for certain styles that have different
inset values. The Imagine style gets its inserts from the
NinePatchImage for instance, which would cause the test to break.

Fix the issue by adding the inset offset when comparing the
background's position.

Task-number: QTBUG-133530
Pick-to: 6.9 6.8
Change-Id: I019630eac0b6704eff795fb81d5b5e39c3362faf
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>