Use leftPadding of parent menus when flipping submenus
QQuickMenuPositioner positions child menus to the right of the menu item
that opened it. In additon, the overlap and rightPadding properties
are used to offset the x position, which styles can use to fine tune how
exactly submenus should be positioned relative to the parent menu item.
The parent menu's padding is used, because the submenu's position is
initially set to be relative to the menu item that opened it. When the
position is offset by the padding, it will position itself to the parent
menu, and not just the parent menu item.

7add1868c0956bca769a33d2974b211c6e533567 introduced the flipping
behavior for menus that have their popupType set to Popup.Window.
Causing menus that don't have enough space to open to the right side, to
instead open on the left side.
The aformentioned patch used the overlap, but forgot to also add the
leftPadding, when flipping submenus.

Also fix the test tst_QQuickMenu::subMenuFlipsPositionWhenOutOfBounds,
which was previously broken on all styles that modified the rightPadding
property of Menu. 6e327f28c7cd05d90881998696da45615 introduced a
temporary fix, in order to unblock CI, and this patch should fix it
properly.

Change-Id: I62c58617dbcf27c097009859213cb78eb59a8008
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 8495325ea10ecd865494132cd460740ff469cf6c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 24b10357a0551fcfd9c887da0534bcf64b02b14d)