CMake: Temporary workaround for missing dependencies in qml_tool
Due to unknown reasons, if the qml_tool target links twice to
the Qt6::Qml target, the qml_tool_autogen target loses almost all
dependencies on other autogen and sync_headers targets, but most
importantly on Qml_sync_headers.

This can cause bad race conditions during the build.

Apply a hacky workaround of removing the duplicate linking to Qt6::Qml
which appears to resolve the issue, and brings back the correct
dependency.

This might be an issue in upstream CMake, but requires further
investigation.

This change is meant to unblock merges to qtdeclarative.

Pick-to: 6.8 6.9
Task-number: QTBUG-133725
Change-Id: Ib6b41d36a19f5d9e1c9fa96aabfbc00f5870990f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>