qmlformat: don't make comments jump around functions
Add more comment anchors around functions to avoid comments from
"jumping through" parenthesis, the "function" keyword or through
commas.

Add missing sourcelocations to AST::FunctionExpression in the parser
for arrow functions. This requires building the FunctionExpression*
early so that it can access the location of the parentheses.
Add the comma location in FormalParameterList in the parser.

This allows qmlformat to properly attach comments to commas, function
keywords, parentheses in lambdas and arrow functions instead of making
the comments "jump" through them.

Fix a domcomparison method to actually ignore parentheses during
qmlformat checks: this is because qmlformat tends to remove parentheses
of lambdas like `(x) => x` that become `x => x`. Now that we actually
save the location of the parentheses, the domcomparison method sees
that the parentheses disappear and complains wrongfully. Therefore,
ignore the newly added parentheses sourcelocations when doing
qmlformat checks. We have plenty of other tests that make sure the
parentheses are the way they should be.

Pick-to: 6.9 6.8
Task-number: QTBUG-123386
Change-Id: I08a3536cfb27fa1cdc2a78cc9849f609a8113a07
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>