tst_qquickmenu: ensure that the menu doesn't open under the mouse
The "resetCurrentIndexUponPopup" test checks that a menu item is not
pre-selected when a menu opens a second time. But will only
be the case if the menu doesn't open directly under the mouse,
since otherwise the mouse will end up selecting the menu item
underneath it.

This patch will therefore make sure that we move the mouse away
before the menu opens, and that we also tell the menu to open
inside the bounds of the parent window rather than at the location
of the mouse, since the latter turns out to be flaky.

Task-number: QTBUG-133530
Pick-to: 6.8 6.9
Change-Id: I5685e9223a6619cc179cc4e8b7e331b0ded11d20
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>