QAccessible: Improve accessibility for QQuickWidget offscreen items
This commit addresses an issue where the accessible interface
(QAccessibleInterface) returned nullptr as the accessible parent for
quick items managed by QQuickWidget. The problem arose because the
accessible path from the application to the quick items correctly passed
through the QQuickWidget, but the reverse path was broken due to the
root item's parent returning the QQuickWidget's window instead.

To resolve this, the patch implements a check in
qaccessiblequickitem.cpp to determine if the item's window is a
QQuickWidgetOffscreenWindow. If so, it uses the associated QQuickWidget
as the accessible parent. Additionally, in qquickwidget.cpp, we set a
"parentWidget" property on the offscreen window when creating its
instance, ensuring that it references its corresponding QQuickWidget.

The test case creates an instance of QQuickWidget inside the QWidget.
Then it verifies that the accessible items can reach to the quick
widget.

Fixes: QTBUG-130116
Change-Id: I431e7e1aba6f7e64a2d5a0c1273ab7b008b962de
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 54310af8f06414d57f22896e4e3f84b3491fe102)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit d0fb323786add149a6cce65b70cc7107661e2c61)