Change background color of the Windows style MenuBar
The background color was hard coded to be white, which prevented it from
adapting to dark mode and high contrast theme changes.

Use button background color instead, since native windows applications
appears to use the same color for both button and menubar backgrounds.

It's also the color role we're assigning to in menubarPalette() in
qwindowstheme.cpp

Task-number: QTBUG-133595
Pick-to: 6.9 6.8
Change-Id: Ic9fdaebb2b38c125b223a5599396d93463ac9cd1
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>