Material: Fix running test_systemTheme with actual Dark system theme
MockPlatformTheme::m_colorScheme was initialized to Unknown, which
translated to System theme, which was handled as not Dark theme
in the test case. This worked, if the actual system theme was Light.

However, when the system theme is Dark, QGuiApplication::styleHints
colorScheme is initialized to Dark as well. In this case, the test
case failed, because the styleHints did not change and the Material
style did not receive theme change event, resulting in test failure.

Pick-to: 6.9
Change-Id: If9bd4a942344b19dffe2b3ffb96faf0c18b9e2b0
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>