vectorimage: Use unhinted font layouts
The whole idea of VectorImage is scalable graphics, and a hinted font
layout is inherently unscalable, as the glyph positions will be rounded
to integers. Therefore, we need to disable hinting on the QFont before
creating the layout.

This also updates the svg manual test to compile and run again, as it
seems to have been broken at some point.

Pick-to: 6.8 6.9
Fixes: QTBUG-133852
Change-Id: Idd63a9fcab51c1f077427ea06ad891975ee73386
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>