ObjectCreator: Fix memory corruption
When we store the QQmlParserStatus for an attached object, we put it
into a vector to call it later. However, the status also contains a
back-reference to the container, so we need reference stability.
That is not given if we use a vector, and the vector needs to
reallocate to grow.

Avoid the issue by using a std::deque, which ensures reference stability
as long as one does not insert in the middle (which we don't). This is
rather suboptimal with MSVC's standard library, but we want a minimal
fix to unblock other patches. Finding a more suitable container can be
deferred to a later point.

Pick-to: 6.9
Fixes: QTBUG-134035
Change-Id: I7c3689ba73060abb47f32a20bc2301b82cf48fa6
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>