tst_qquickpointerhandler: fix warning about mismatched enums; ctor casts
The GrabTransition enum has no "none" value: there's no reason we need
one for public API. But this test records events and transitions, and
needs a way to record that no grab transition happened. It's just an int
anyway, so QCOMPARE_EVENT simply casts it before comparison; and in any
ternary-if expression using NoGrab, we cast both arms to int.

The warning was

warning: enumerated mismatch in conditional expression: ‘<unnamed enum>’ vs
  ‘QPointingDevice::GrabTransition’ [-Wenum-compare]
  549 | QEventPoint::State::Pressed, itemIgnoreMouse ? NoGrab : QPointingDevice::GrabExclusive);
      |                                              ^

Next, there is -Wint-in-bool-context: fixed by surrounding the ternary
if's with parentheses.

Replace C-style with constructor-style casts while we're at it.

Change-Id: I0796abc8e19c1842a6b2f4e343d566ff0cc226e6
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>