QmlModels: Move poolTime out of QQmlDelegateModelItem
The time it has spent in the pool is only relevant while it actually is
in the pool. There is no reason to waste the space for the poolTime
while it's not.

Furthermore, we don't need QList's implicit sharing for the pool. Use
std::vector instead. We do, however, need to check the pool for
overflows.

Change-Id: Ib9d545367c41dacf73ae4bc391af7928fa67ea2f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>