Material: Fix FloatingPlaceholderText X position
In cb7eb152204e206539f307a9556eea43c589f026, we stopped
using the leftPadding value for FloatingPlaceholderText.

As a result, the text is now misaligned with the placeholder.

This commit restores the FloatingPlaceholderText X position to
leftPadding. If leftPadding is not set, it defaults to
Material.textFieldHorizontalPadding.

To prevent regressions (such as QTBUG-120149), a new X position
animation has been added to FloatingPlaceholderText.

Fixes: QTBUG-133492
Pick-to: 6.9 6.8
Change-Id: I5c80dcedd7c2beec0891b524fc55388fa5456f2a
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>