CMake: Avoid leaking Private dependencies for style plugins
Use qt_internal_extend_target instead of target_link_libraries for
the Qt Quick style plugins.
This wraps their private module dependencies in BUILD_INTERFACE
genexes. It avoids leaking the deps to consumers of the style plugins
in a static build, where the private modules are not in scope by
default.

Pick-to: 6.8 6.9
Change-Id: Iac4b9d97c112fa829a60170d7fdad60f97e43f1f
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>