QQmlSA: Remove PassManager::{element|property}Passes() from public API
It complicates things internally and doesn't add much value to the user.
They can still keep track of which passes they have created manually.

QQmlSA is still in technical preview.

[ChangeLog][QML][QQmlSA] PassManager will no longer expose the element
and property passes that it holds. They are only meant to be used inside
the PassManager itself. elementPasses() and propertyPasses() were
removed.

Pick-to: 6.9
Change-Id: Ibc748309f142794a27266e68ad9d0ce302b51213
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>