TableView: emit commit on editor loss of focus, unless Qt::Key_Escape
As it stood, we would only commit changes in the edit delegate when
the user pressed Enter, Return, or Tab/BackTab. The result was that
the changes would be not be committed (and hence lost) if the user
clicked somewhere else in the UI to transfer focus.

This behavior is basically wrong. Instead, it should be the
opposite, meaning that we should _always_ commit the changes unless
the user explicitly cancelled editing by pressed Escape.
This behavior is also verified by testing macOS "Numbers", and
Google "Sheets" (both spreadsheet apps).

This patch will therefore change the faulty behavior, so that
we commit the changes whenever the edit item loses focus, unless
it happened because the user hit Escape.

[ChangeLog][Quick][TableView] The edit delegate will now emit the
onCommit signal when it loses focus, unless it was closed from
Qt::Key_Escape.

Fixes: QTBUG-134247
Pick-to: 6.8
Change-Id: I1afab7bce9ba159330203d5a8d2b1af57ce9875b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 48cb120c5f0bb5ce093b46bfe63751b4bd01caf9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>