Fix flaky test tst_qquickColorDialogImpl::dialogCanMoveBetweenWindows
tst_qquickColorDialogImpl::dialogCanMoveBetweenWindows() is flaky where
in some cases the quickDialog cannot be close properly as no signal is
emitted when clicking the button.

When it is failing, the re-parenting of the quickDialog is done
properly, but the quickDialog is not rendered. The button cannot then
be clicked to close the dialog.

Close the popupWindow() and open it, before closing it with
CLOSE_DIALOG("Ok").

Fix flaky tst_qquickColorDialogImpl::dialogCanMoveBetweenWindows on
Opensuse

Pick-to: 6.9 6.8 6.5
Fixes: QTBUG-134043
Change-Id: Ib60ecd745e0407fa67ba2dfcaf8e445dc96a0f06
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>