Auto-depend on QtQuick when linking against QtQuick
When creating a QML module and deriving from a C++ type exposed by
another module, it is necessary to add the originating module to
DEPENDENCIES, lest tooling won't find it.

In most cases, this is not done, and the fact that at runtime everything
still works is a cause of unhappiness and complaints against tooling.

We can't fix this in general (without introducing an additional
qml_link_module function, which would combine target_link_libraries with
setting DEPENDENCIES).

However, for the most common case, which is deriving from a type from
QtQuick, notably QQuickItem, we can attempt a heuristic: If someone
explicitly links against QtQuick in a QML module, we assume that they
actually want to have QtQuick in dependencies.

[ChangeLog][QML CMake API] If a QML module target is linked against
Qt6::Quick, QtQuick is automatically added to its QML dependencies. This
avoids tooling errors when e.g. a QQuickItem derived type is exported by
the module.

Pick-to: 6.9 6.8
Change-Id: I6132e9c49df5b7f41cff58182768e85c68734bc7
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>