QtQuickView statuslistener CI tests
This test adds basic loading status tests for QtQuickView on Android.
The test uses a primary QtQuickView which has the TestCases, and a
Communicator QML_ELEMENT C++ class to load other QML files into a
secondary QtQuickView contained in the hosting layout. The hosting
activity listens to QtQuickView status changed signals, and sends those
directly to the Communicator object, which emits a signal with the
status information to the TestCases.

Task-number: QTBUG-133439
Change-Id: I01a3a1dc3124036d8290a248b9b2d1bb4010033d
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>