Change the way we do color animations
We need a generalized way of doing animations to simplify
adding animations for other properties. In order to be
as expressive as possible we need to just pass the
animation info on to the generator and let it handle it.
If there are multiple animations affecting the same property
at the same time, we just instantiate them in Qt Quick
and let its engine handle the conflict.

This means that the only way to implement special SMIL
conflict behavior such as add/replace is to create specialized
animation classes in Qt Quick. We can do this at a later
stage if it becomes a priority.

The main use case of having multiple animations affecting
a property at different times should be covered by this
implementation, it should be easy to expand to new types,
and it should be easy to add stuff like easing curves to
it.

Change-Id: I927ea2cef214a300ae8fae3f76ee87fb5d505b3d
Reviewed-by: Hatem ElKharashy <hatem.elkharashy@qt.io>