QQuickItemView: Fix model creation
In setDelegate(), we want to create the model if it doesn't exist, not
if it's not ours. Otherwise we might overwrite a perfectly fine model
set by user code when creating a delegate.

Also warn when we cannot retain an explicitly set delegate because an
incompatible model was passed.

Pick-to: 6.8
Change-Id: I43635d4cc2fcd0f539e1cfdf742a7d941323c2fe
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit cb59f99d0bfde35b8401b6249e5696341f3f6f98)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>