Expose delegateModelAccess from QQuickItemView
It mirrors the same property from any internal delegate model. If the
internal model is not a delegate model, the value is "Qt5ReadWrite".

[ChangeLog][QtQuick] ListView and GridView now have a new property
delegateModelAccess. Setting it to DelegateModel.ReadWrite allows you to
write values into the model via required properties just as you could
with context properties.

Task-number: QTBUG-132420
Change-Id: I72e27fd737ef2ef25da067e4192b58aad9084143
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>