AbstractDialog: add property popupType
18c4bf827dabea735435887f84ad1bea4136b1ea made all non-native dialogs
in QtQuick.Dialogs use `Window` as their popupType, in order to make
the dialogs appear as top-level windows. However, there are some users
that wish to use these non-native dialogs without them being top-level
windows, like they were before 6.8. This was unfortunately not possible,
since there weren't any API to customize the popupType for
QtQuick.Dialogs.

We were reluctant at first, to provide properties that only customize
the non-native dialogs, without having any effect when native dialogs
are used. However, there needs to exist some way for the popupType to be
reverted back to `Item`, and thus, it makes sense to have a property for
it.

Fixes: QTBUG-134037
Change-Id: Idcfe4d3960c7002566f344ae32422c9cb7ce7631
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>