test: Improve robustness of tst_QQuickText::baselineOffset()
The scaled data sets for the baselineOffset test basically
implements another version of the font-fitting algorithm and then
checks that the one in Qt Quick gives the same results as this.

But there was one crucial difference: The algorithm that provides
the expected value has the maximum font size at 10000, so it will
try to find the best fit between 10000 and 0. But the maximum
size in the "actual" result was set to 64, so it would try to
find a match between 0 and 65.

Therefore the two algorithms might end up with different fitting
values that were within the margin of 0.01.

In one particular case, on macOS, it seems that the point size
of 64 just happened to fit exactly inside the width of 200.
Due to some change in Harfbuzz 10.3.0 (most likely the AAT
optimizations), it became possible to find a better match, but
this appears to have been a ticking time bomb.

Pick-to: 6.5 5.15
Change-Id: I87188cbfd9070edc7b9c036f18d005ab9dd24026
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 4c3584e22563e268e7beb97d592612d8254cfb87)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit e5327ff188d4692078429fd4516b9d2479192b44)
(cherry picked from commit 1ec4344467aae4ed90cf9708431c16f59afd1729)