Control: fix regression where hoverEnabled defaulted to false
Now (after 83d90845527799c78366c2b89e9c14bc35e88695) that
QQuickContentItem's parent is the content control rather
than QQuickRootItem, we need to ensure that the control has the
correct hoverEnabled value. If we don't do this (get the value
from the style hints), it will use QQuickItem's default value of false
and some controls won't be hoverable when they should be.

Fixes: QTBUG-133886
Change-Id: Ib0912acf7cdf513bf004837bc4ede970decf23d8
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>