CMake: Fix QuickControls2 QML tests test to pass in a static build
The tst_basic test was failing when building against a static Qt build
because the QML import scanner was not able to find the QML files that
are outside the current source directory.

Fix it by adding the test's data directory to the
QT_QML_IMPORT_SCANNER_EXTRA_ROOT_PATHS property.

Pick-to: 6.8 6.9
Fixes: QTBUG-131813
Change-Id: If52375708b17e249fb2998ec8f92169f267e1980
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>