Set fix position tst_QQuickHoverHandler::window to fix flakiness
tst_QQuickHoverHandler::window() is flaky on Opensuse, where the cursor
is not hovering the window when it is failing.

Set fix position to the window for the cursor to always hover at the
right position.

Unblacklist tst_QQuickHoverHandler::window() on Opensuse as it is no
longer flaky.

Fixes: QTBUG-122405
Pick-to: 6.8 6.5
Change-Id: I0c78e80d19c59e95d58765ce2f6e201121e14236
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit a9095af4d06f954d12f42e1c7c2745d91875398c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>