Remove request for activation tst_QQuickApplication::state
tst_QQuickApplication::state is flaky on openSuSE where when changing
the focus to nullptr, it remains on the window. This is due to some
asynchronicity when requestActivate is call, requesting a second
activation when the focus should be changed to nullptr.

Do not call requestActivate() on the window when the platform uses xcb

Unblacklist tst_QQuickapplication::state() on openSuSE as no longer
flaky.

Fixes: QTBUG-122031
Pick-to: 6.8 6.5
Change-Id: I66ebc02bc44c6ba27d13f5eb5f026006ddf13263
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 11404fd4ea6d1c61f49b39afaa1951961809c2f5)