Remove request for activation tst_QQuicApplication::active
tst_QQuickApplication::active is flaky on openSuSE where when changing
the focus to nullptr, it remains on the window. This is due to some
asynchronicity when requestActivate is call, requesting a second
activation when the focus should be changed to nullptr.

Do not call requestActivate() on the window when the platform uses xcb

Fixes: QTBUG-75215
Fixes: QTBUG-123550
Pick-to: 6.8 6.5
Change-Id: I77216767f021abed09c888a17e1a580900044a18
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>
(cherry picked from commit 6ba26f74df58bbbf102480a4d309624480c7bcf3)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>