QQuickRepeater: Fix model creation
In setDelegate(), we want to create the model if it doesn't exist, not
if it's not ours. Otherwise we might overwrite a perfectly fine model
set by user code when creating a delegate.

Also warn when we cannot retain an explicitly set delegate because an
incompatible model was passed.

Pick-to: 6.8
Change-Id: I11cf3e30d8966da6489882b263d56af5df8140b2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit 50c616c5605a5a223d058d69336dc82d48040e98)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>