QQuickDeliveryAgentPrivate: fix performance regression in eventTargets()
a89a34463bff3d00b96732fd3ebf84116a645248 introduced a predicate
parameter to eventTargets. The problem is that it is passed by
std::function copy. Since eventTargets is called recursively it can
easily, in a heavy scene, lead to hundreds of allocation + deallocations
for each single event handling.

Pass the predicate as a (qxp::)function_ref to fix this.

Since function_ref has reference semantics, we cannot inititalize it
with nullptr anymore. So define a noop predicate for the
contextMenuTargets() case.

Pick-to: 6.9
Change-Id: I6a3162b9b76b197bba07c1725ac7768c8883818b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>