qmlformat: Remove PendingSourceLocation::toUpdate
There are many places in QmlFormat (DOM in general) relying a lot on the obscure side-effects. toUpdate member of PendingSourceLocation is one of these.

Typically the data behind it gets re-assigned under not so clear conditions in the `LineWriter::commitLine`.
This commit removes this member and simplifies all the upstream usages of it.

Unfortunately because the details are somewhat fuzzy and requires more research, this change is mildly risky, relying mostly on the existing tests, which seem to be passing fine.
Therefore it's open to the discussion whether it's OK to take this risk or better not.

Change-Id: Ia3ea68cd0afc44f1e8dbb35163ad29211c884b5b
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>