QQuickPopupItem: Set tabFence to be true by default
The tabFence property should not be dependent on popup modality as it
was introduced in 611422b5968ceac6cfba034116a3d1e107d8e2df. The popup
item should by default act as a tabFence, meaning that it should not
allow tab focus navigation in and out of a popup, no matter its
modality.
However, certain kinds of popups, such as the drawer, where it
makes sense to tab focus outside, should override the tabFence
property to be false. In that case, keep the old behavior that makes
this dependent on the modality.

Reverts 611422b5968ceac6cfba034116a3d1e107d8e2df and
e9555f6de625a2fffdfb032a7c293f842b269f9a

[ChangeLog][QtQuick][Controls][Popup] The popupItem always acts as
a tab fence, preventing tab focus navigation in and out of the popup, except when the popup is a Drawer, in which case this behavior depends on the drawer's modality.

Fixes: QTBUG-132763
Change-Id: I99d250333c98354429943ca6b3267482b2152b47
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>