qmlformat: remove LineWriter::startSourceLocation()
The previous patch gives some hints that the only one more usage of startSourceLocation() can probably be eliminated, as well as the whole function.
However, as mentioned in the previous patch, the area of formatter is quite fragile and messy, there are some unknown risks (because of lack of knowledge) associated with this patch.

Change-Id: I3a1184ec9dbc31762a12a34ba602f17049702522
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>