QQmlPluginImporter: refactor makePlugins()
This is also a preparation to make a clearer separation between querying, filtering and matching of qml static plugins.

This patch attempts to encapsulate the process of verifying whether a static plugin is a qml plugin and if so, extracting metadataURI-s. Against which the plugin will be matched at a later point (populatePluginDataVector).
Making the functionality of makePlugins more straightforward :
1. get all static plugins
2. filter out non qml plugins

At a later point tryConvertToStaticPluginData we'll be renamed and refactored to tryExtractQmlPluginURIs. As well as makePlugins to staticQmlPlugins.

Change-Id: I8b43d6c597358daf258ad19817f9389d69131b2b
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>